Fold enhanced check page tests into the controller test
The standalone enhanced-check.test.js duplicated the controller's mock
and request setup to exercise the feature-gated check page. The routing
decision lives in the controller, so the tests belong alongside it: this
moves all 14 tests into controller.test.js as a top-level describe that
turns the feature flag on, reusing the existing mocks and request
helper, and deletes the standalone file. Once the flag is permanent and
the legacy page is removed, these assertions stay put rather than being
orphaned.
